addressing reviewer's comment
Removed Netfx and Netcore files for SqlSer and added Netfx if endifs in common file
